hund (hund v0.1.1)

Types

assertion/0

-type assertion() :: #saml_assertion{}.

authn/0

-type authn() :: #saml_authn{}.

authn_class/0

-type authn_class() ::
    password | password_protected_transport | internet_protocol | internet_protocol_password |
    mobile_one_factor_contract | mobile_two_factor_contract | previous_session | unspecified.

authnreq/0

-type authnreq() :: #saml_authnreq{}.

condition/0

-type condition() :: #saml_condition{}.

contact/0

-type contact() :: #saml_contact{}.

datetime/0

-type datetime() :: string() | binary().

idp_metadata/0

-type idp_metadata() :: #saml_idp_metadata{}.

localized_string/0

-type localized_string() :: string() | [{Locale :: atom(), LocalString :: string()}].

logout_request/0

-type logout_request() :: #saml_logout_request{}.

logout_response/0

-type logout_response() :: #saml_logout_response{}.

name_format/0

-type name_format() :: email | x509 | windows | krb | persistent | transient | unknown.

org/0

-type org() :: #saml_org{}.

saml_record/0

-type saml_record() ::
    contact() |
    org() |
    idp_metadata() |
    sp_metadata() |
    subject() |
    assertion() |
    authn() |
    authnreq() |
    logout_request() |
    logout_response().

sp_metadata/0

-type sp_metadata() :: #saml_sp_metadata{}.

status_code/0

-type status_code() ::
    success | request_error | response_error | bad_version | authn_failed | bad_attr | denied |
    bad_binding | unknown.

subject/0

-type subject() :: #saml_subject{}.

subject_method/0

-type subject_method() :: bearer | holder_of_key | sender_vouches.

version/0

-type version() :: string().

Functions

check_dupe_ets(A, Digest)

-spec check_dupe_ets(esaml:assertion(), Digest :: binary()) -> ok | {error, duplicate_assertion}.

Checks for a duplicate assertion using ETS tables in memory on all available nodes.
This is a helper to be used as a DuplicateFun with hund_sp:validate_assertion/3. If you aren't using standard erlang distribution for your app, you probably don't want to use this.

convert_fingerprints(FPs)

-spec convert_fingerprints([string() | binary()]) -> [binary()].

Converts various ascii hex/base64 fingerprint formats to binary

date_to_saml(Date)

-spec date_to_saml(calendar:date()) -> string() | binary().

datetime_to_saml(Time)

-spec datetime_to_saml(calendar:datetime()) -> datetime().

Converts a calendar:datetime() into SAML time string

import_certificate(EncodedCert, Identifier)

-spec import_certificate(EncodedCert :: string(), Identifier :: term()) -> binary().

import_private_key(EncodedKey, Identifier)

-spec import_private_key(EncodedKey :: string(), Identifier :: term()) -> #'RSAPrivateKey'{}.

load_certificate(Path)

-spec load_certificate(Path :: string()) -> binary().

load_metadata(Url)

-spec load_metadata(Url :: string()) -> esaml:idp_metadata().

Reads IDP metadata from a URL (or ETS memory cache)

load_metadata(Url, Fingerprints)

-spec load_metadata(Url :: string(), Fingerprints :: [string() | binary()]) -> esaml:idp_metadata().

Reads IDP metadata from a URL (or ETS memory cache) and validates the signature

load_private_key(Path)

-spec load_private_key(Path :: string()) -> #'RSAPrivateKey'{}.

Loads a private key from a file on disk (or ETS memory cache)

map_authn_class(AuthnClass)

-spec map_authn_class(AuthnClass :: string()) -> authn_class().

map_if(K)

-spec map_if(term()) -> [term()].

map_if(Key, List)

-spec map_if(atom(), term()) -> [term()].

nameid_map(S)

-spec nameid_map(string()) -> name_format().

rev_map_authn_class(Context)

-spec rev_map_authn_class(Context :: atom()) -> string().

rev_nameid_map(_)

-spec rev_nameid_map(atom()) -> string().

rev_status_code_map(Status)

-spec rev_status_code_map(status_code() | atom()) -> string().

rev_subject_method_map(_)

-spec rev_subject_method_map(subject_method()) -> string().

saml_to_datetime(Stamp)

-spec saml_to_datetime(esaml:datetime()) -> calendar:datetime().

Converts a SAML time string into a calendar:datetime()
Inverse of datetime_to_saml/1

status_code_map(Urn)

-spec status_code_map(string()) -> status_code() | atom().

threaduntil(Rest, InitAcc)

-spec threaduntil([fun((Acc :: term()) -> {error, term()} | {stop, term()} | term())],
            InitAcc :: term()) ->
               {error, term()} | {ok, term()}.

unique_id()